Fixed a bug where adding a new item to a reading list wasn't adding it at the end.
